services/ai-chat/pyproject.toml
[project]
name = "ai-chat"
version = "0.1.0"
description = "AI Chat FastAPI Service"
requires-python = ">=3.11"
dependencies = [
    "fastapi>=0.115.0",
    "uvicorn[standard]>=0.31.0",
    "pydantic-settings>=2.5.0",
    "pyjwt[crypto]>=2.9.0",
    "cryptography>=41.0.0",
    "httpx>=0.27.0",
    "semantic-kernel>=1.0.0",
]

[tool.uv]
dev-dependencies = []
